Init pyodide/webR engine based on runApp arg
A new argument `engine` is added to runApp() that selects webR or
Pyodide as the underlying engine to use.

The main app selects a React hook `useWebR()` or `usePyodide()` based on
this argument.

The proxy handle variable `proxyHandle` is typed as a union,
`ProxyHandle = PyodideProxyHandle | WebRProxyHandle`, and the property
`engine` is added as a way to discriminate between which Wasm engine is
currently loaded.

Hooks relying on Pyodide are disabled through early return when the
`proxyHandle.engine` property is not equal to "pyodide".