Only insert COI headers if requested in URL param
With this commit the Service Worker will only inject the HTTP
headers required for Cross-Origin Isolation if the URL or referrer
URL contains the parameter `coi=1`.

Now, Shinylive for Python is not served with COI headers injected
by default. This fixes the Map example when running under Pyodide.

If an app using the webR engine is run, and the page is not
Cross-Origin Isolated, the parameter `coi=1` is added to the current
location URL and the page refreshed. With this, using webR without
the required COI headers from the server is handled automatically.